Fixed error by changing from double to int in kap6/ovning6_9.c.

+++ kap6/ovning6_9.c
@@ -2,7 +2,7 @@
 
 double read_double(void);
 int read_int(void);
-double upphojt_till(double,double);
+double upphojt_till(double,int);
 
 int main(void)
 {
